renaming controller to document controller
