+++ src/document-utils-api/src/main/java/ca/bc/gov/open/jag/documentutils/api/DocumentController.java
@@ -1,6 +1,4 @@
-package ca.bc.gov.open.jag.documentutils.controller;
-
-import javax.validation.Valid;
+package ca.bc.gov.open.jag.documentutils.api;
 
 import ca.bc.gov.open.jag.documentutils.exception.MergeException;
 import ca.bc.gov.open.jag.documentutils.model.DocMergeRequest;
@@ -9,17 +7,14 @@ import ca.bc.gov.open.jag.documentutils.model.JSONResponse;
 import ca.bc.gov.open.jag.documentutils.service.MergeService;
 import ca.bc.gov.open.jag.documentutils.utils.DocMergeConstants;
 import ca.bc.gov.open.jag.documentutils.utils.DocMergeUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
+import javax.validation.Valid;
 import java.text.MessageFormat;
 
 /**
@@ -32,28 +27,29 @@ import java.text.MessageFormat;
 
 @RestController
 @Validated
-public class MergeController {
+@RequestMapping("document")
+public class DocumentController {
 
 	private final MergeService mergeService;
 	
-	private final Logger logger = LoggerFactory.getLogger(MergeController.class);
+	private final Logger logger = LoggerFactory.getLogger(DocumentController.class);
 
-	public MergeController(MergeService mergeService) {
+	public DocumentController(MergeService mergeService) {
 		this.mergeService = mergeService;
 	}
 
-	@PostMapping(value = {"/merge/{correlationId}" }, 
+	@PostMapping(value = {"/merge" },
 			consumes = DocMergeConstants.JSON_CONTENT,
 			produces = DocMergeConstants.JSON_CONTENT)
 	public ResponseEntity<JSONResponse<DocMergeResponse>> mergeDocumentPost(
-			@PathVariable(value = "correlationId", required = true) String correlationId, 
+			@RequestHeader(value = "X-TransactionId", required = true) String transactionId,
 			@Valid @RequestBody(required = true) DocMergeRequest request)  {
 		
 		logger.info("Starting merge process...");
 
 		try {
 			
-			DocMergeResponse mergResp = mergeService.mergePDFDocuments(request, correlationId);
+			DocMergeResponse mergResp = mergeService.mergePDFDocuments(request, transactionId);
 			JSONResponse<DocMergeResponse> resp = new JSONResponse<>(mergResp);
 			logger.info("Merge process complete.");
 			return new ResponseEntity<>(resp, HttpStatus.OK);
@@ -62,7 +58,7 @@ public class MergeController {
 
 			logger.error(MessageFormat.format("Document Merge encountered an error: {0}", e.getMessage()), e);
 			return new ResponseEntity<>(
-					DocMergeUtils.buildErrorResponse(String.format(DocMergeConstants.NOT_PROCESSED_ERROR, correlationId), 500),
+					DocMergeUtils.buildErrorResponse(String.format(DocMergeConstants.NOT_PROCESSED_ERROR, transactionId), 500),
 					HttpStatus.INTERNAL_SERVER_ERROR);
 
 		}
